hw_status/hw_status/requirements.txt
fastapi>=0.115.12
starlette>=0.46.2
uvicorn>=0.34.0
alembic>=1.14.1
tenacity>=9.0.0
dnspython>=2.7.0
pydantic>=2.10.6
pydantic_settings>=2.7.1
dnspython>=2.7.0
flake8>=7.2.0
simplejson>=3.19.3
httpx>=0.28.1
python-slugify>=8.0.4
python-escpos>=1.0.9
pyserial>=3.5
pillow>=11.2.1
qrcode>=8.2
python-barcode>=0.15.1
jinja2>=3.1.6